+++ app/routes/auth.py
@@ -13,7 +13,7 @@ from datetime import datetime, timedelta, timezone
 import re
 from app import db, jwt
 from app.models.user import User
-from app.utils.validators import validate_email, validate_password, error_response
+from app.utils.validators import validate_email, validate_password, error_response, validate_password_complexity
 from werkzeug.exceptions import BadRequest
 from sqlalchemy.exc import IntegrityError
 import time
@@ -313,31 +313,3 @@ def change_password():
     except Exception as e:
         current_app.logger.error(f"Password change error: {str(e)}")
         return error_response("Failed to change password", 500)
-
-
-def validate_password_complexity(password):
-    """
-    Validate that a password meets complexity requirements
-    - At least 8 characters long, maximum 128 characters
-    - Contains uppercase letter
-    - Contains lowercase letter
-    - Contains a number
-    - Contains a special character
-    """
-    if not isinstance(password, str) or not password:
-        return False
-
-    # For testing convenience, accept simple passwords in test mode
-    if current_app.config.get("TESTING"):
-        return len(password) >= 8
-
-    if len(password) < 8 or len(password) > 128:
-        return False
-
-    # Check for at least one uppercase, lowercase, digit and special character
-    has_uppercase = any(c.isupper() for c in password)
-    has_lowercase = any(c.islower() for c in password)
-    has_digit = any(c.isdigit() for c in password)
-    has_special = any(not c.isalnum() for c in password)
-
-    return has_uppercase and has_lowercase and has_digit and has_special
